CosmiXs

Documentation for CosmiXs.

CosmiXs.dNdE

CosmiXs.dNdE — Method

dNdE(mDM::Energy, K::Energy; final_particles="Gammas", primary_channel="b")

Return the differential flux of final_particles from the annihilation of dark matter into primary_channel particles, given the energy of final particles K and the dark matter mass mDM.

Available final_particles are: AntiProtons, Gammas, NuEl, NuMu, NuTau, Positrons.

Available primary_channel are:

eL, eR, e,

muL, muR, mu,

tauL, tauR, tau,

nu_e, nu_mu, nu_tau,

u, d, s, c, b, t,

a, g,

W, WL, WT,

Z, ZL, ZT,

H,

aZ,

HZ